Create new folders and move corresponding files there.
