+++ app/src/main/java/ru/androidacademy/gooseeye/models/ArtistInfo.kt
@@ -1,3 +1,3 @@
-package ru.androidacademy.gooseeye
+package ru.androidacademy.gooseeye.models
 
 data class ArtistInfo(var artistPhoto: Int, var artistName: String)
