+++ app/src/main/res/drawable/rating_bar_details.xml
@@ -3,12 +3,12 @@
 
     <item
         android:id="@android:id/background"
-        android:drawable="@drawable/ic_star_icon_gray" />
+        android:drawable="@drawable/star_empty_details" />
     <item
         android:id="@android:id/secondaryProgress"
-        android:drawable="@drawable/ic_star_icon_gray" />
+        android:drawable="@drawable/star_empty_details" />
     <item
         android:id="@android:id/progress"
-        android:drawable="@drawable/ic_star_icon" />
+        android:drawable="@drawable/star_filled_details" />
 
 </layer-list>
