+++ app/src/main/res/layout/activity_movie_details.xml
@@ -82,10 +82,10 @@
 
         <TextView
             android:id="@+id/tv_name"
-            style="@style/name_style"
+            style="@style/name_details_style"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
-            android:text="@string/name_text"
+            android:text="@string/name_avengers"
             app:fontFamily="sans-serif-black"
             app:layout_constraintLeft_toRightOf="@id/guideline_margin_start"
             app:layout_constraintRight_toLeftOf="@id/guideline_margin_end"
@@ -93,7 +93,7 @@
 
         <TextView
             android:id="@+id/tv_tag_line"
-            style="@style/tag_line_style"
+            style="@style/tag_line_style_details"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:layout_marginTop="4dp"
@@ -104,11 +104,11 @@
 
         <ru.androidacademy.gooseeye.RatingBarSvg
             android:id="@+id/ratingBar"
-            style="@style/custom_rating_bar_style"
+            style="@style/rating_bar_details_style"
             android:layout_width="wrap_content"
             android:layout_height="14dp"
             android:numStars="5"
-            android:progressDrawable="@drawable/custom_rating_bar"
+            android:progressDrawable="@drawable/rating_bar_details"
             android:rating="4"
             android:stepSize="0.01"
             android:layout_marginTop="7dp"
@@ -143,7 +143,7 @@
 
         <TextView
             android:id="@+id/tv_story_line_details"
-            style="@style/story_line_details_style"
+            style="@style/white_14sp"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:layout_marginTop="4dp"
